requirements.txt
flask
torch
torchvision
timm
transformers
pillow
requests
numpy